ai/server/auth.py
import jwt
import datetime
from flask import request, jsonify
from functools import wraps

from dotenv import load_dotenv
import os

# Load environment variables from the .env file (if present)
load_dotenv()

# Secret key for JWT encoding/decoding
SECRET_KEY = os.getenv('SECRET_KEY', 'test-key')

# JWT Token Expiry Duration
TOKEN_EXPIRY_MINUTES = 30

def generate_token(user_id):
    payload = {
        'user_id': user_id,
        'exp': datetime.datetime.utcnow() + datetime.timedelta(minutes=TOKEN_EXPIRY_MINUTES)
    }
    token = jwt.encode(payload, SECRET_KEY, algorithm='HS256')
    return token

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None

        # Token passed in headers
        if 'Authorization' in request.headers:
            token = request.headers['Authorization'].split(" ")[-1]

        if not token:
            return jsonify({'message': 'Token is missing!'}), 401

        try:
            data = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
            current_user = data['user_id']
        except jwt.ExpiredSignatureError:
            return jsonify({'message': 'Token has expired!'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'message': 'Invalid token!'}), 401

        return f(current_user, *args, **kwargs)

    return decorated
import jwt
import datetime
from flask import request, jsonify
from functools import wraps

from dotenv import load_dotenv
import os

# Load environment variables from the .env file (if present)
load_dotenv()

# Secret key for JWT encoding/decoding
SECRET_KEY = os.getenv('SECRET_KEY', 'test-key')

# JWT Token Expiry Duration
TOKEN_EXPIRY_MINUTES = 30

def generate_token(user_id):
    payload = {
        'user_id': user_id,
        'exp': datetime.datetime.utcnow() + datetime.timedelta(minutes=TOKEN_EXPIRY_MINUTES)
    }
    token = jwt.encode(payload, SECRET_KEY, algorithm='HS256')
    return token

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None

        # Token passed in headers
        if 'Authorization' in request.headers:
            token = request.headers['Authorization'].split(" ")[-1]

        if not token:
            return jsonify({'message': 'Token is missing!'}), 401

        try:
            data = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
            current_user = data['user_id']
        except jwt.ExpiredSignatureError:
            return jsonify({'message': 'Token has expired!'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'message': 'Invalid token!'}), 401

        return f(current_user, *args, **kwargs)

    return decorated
